Moved request model test to single file
Instead of using 2 with TestSource
